Update style.py
if the train2014 zp file is not extracted or extracted without setup.sh, the script will run to end without throwing an error.
In my case i unzipped train2014 and it extracted files the second level ie. train2014/train2014/.
we can assert that files are not null in _get_files method.